Consolidate transaction test environments
