+++ server/src/test/java/io/spine/server/entity/given/tx/TxTestProcessManager.java
@@ -18,16 +18,15 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.server.projection.given;
+package io.spine.server.entity.given.tx;
 
 import com.google.common.collect.ImmutableList;
 import com.google.protobuf.Message;
-import io.spine.core.Subscribe;
-import io.spine.server.projection.Projection;
-import io.spine.test.projection.Project;
-import io.spine.test.projection.ProjectId;
-import io.spine.test.projection.event.PrjProjectCreated;
-import io.spine.test.projection.event.PrjTaskAdded;
+import io.spine.server.entity.given.tx.event.TxProjectCreated;
+import io.spine.server.entity.given.tx.event.TxTaskAdded;
+import io.spine.server.event.React;
+import io.spine.server.model.Nothing;
+import io.spine.server.procman.ProcessManager;
 import io.spine.validate.ConstraintViolation;
 import org.checkerframework.checker.nullness.qual.Nullable;
 
@@ -36,22 +35,23 @@ import java.util.List;
 import static com.google.common.collect.Lists.newLinkedList;
 
 /**
- * Test environment projection for {@link io.spine.server.projection.ProjectionTransactionTest}.
+ * Test environment Process Manager for {@link io.spine.server.procman.PmTransactionTest}.
  */
-public class TxTestProjection
-        extends Projection<ProjectId, Project, Project.Builder> {
+public class TxTestProcessManager
+        extends ProcessManager<ProjectId, Project, Project.Builder> {
 
     private final List<Message> receivedEvents = newLinkedList();
-    private final List<ConstraintViolation> violations;
+    private final @Nullable List<ConstraintViolation> violations;
 
-    public TxTestProjection(ProjectId id) {
+    public TxTestProcessManager(ProjectId id) {
         this(id, null);
     }
 
-    public TxTestProjection(ProjectId id,
-                            @Nullable ImmutableList<ConstraintViolation> violations) {
+    public TxTestProcessManager(ProjectId id, @Nullable List<ConstraintViolation> violations) {
         super(id);
-        this.violations = violations;
+        this.violations = violations == null
+                          ? null
+                          : ImmutableList.copyOf(violations);
     }
 
     @Override
@@ -62,24 +62,25 @@ public class TxTestProjection
         return super.checkEntityState(newState);
     }
 
-    @Subscribe
-    void event(PrjProjectCreated event) {
+    @React
+    Nothing event(TxProjectCreated event) {
         receivedEvents.add(event);
         Project newState = Project.newBuilder(state())
                                   .setId(event.getProjectId())
                                   .build();
         builder().mergeFrom(newState);
+        return nothing();
     }
 
     /**
-     * Always throws {@code RuntimeException} to emulate an error in
-     * a subscribing method of a Projection.
+     * Always throws {@code RuntimeException} to emulate the case of an error in
+     * a reacting method of a Process Manager.
      *
-     * @see io.spine.server.projection.ProjectionTransactionTest#createEventThatFailsInHandler
+     * @see io.spine.server.procman.PmTransactionTest#createEventThatFailsInHandler()
      */
-    @Subscribe
-    void event(PrjTaskAdded e) {
-        throw new RuntimeException("that tests the projection tx behaviour");
+    @React
+    Nothing event(TxTaskAdded event) {
+        throw new RuntimeException("that tests the tx behaviour for process manager");
     }
 
     public List<Message> receivedEvents() {
