+++ server/src/test/java/io/spine/server/entity/given/tx/TxTestProjection.java
@@ -18,17 +18,14 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.server.procman.given;
+package io.spine.server.entity.given.tx;
 
 import com.google.common.collect.ImmutableList;
 import com.google.protobuf.Message;
-import io.spine.server.event.React;
-import io.spine.server.model.Nothing;
-import io.spine.server.procman.ProcessManager;
-import io.spine.test.procman.Project;
-import io.spine.test.procman.ProjectId;
-import io.spine.test.procman.event.PmProjectCreated;
-import io.spine.test.procman.event.PmTaskAdded;
+import io.spine.core.Subscribe;
+import io.spine.server.entity.given.tx.event.TxProjectCreated;
+import io.spine.server.entity.given.tx.event.TxTaskAdded;
+import io.spine.server.projection.Projection;
 import io.spine.validate.ConstraintViolation;
 import org.checkerframework.checker.nullness.qual.Nullable;
 
@@ -37,23 +34,22 @@ import java.util.List;
 import static com.google.common.collect.Lists.newLinkedList;
 
 /**
- * Test environment Process Manager for {@link io.spine.server.procman.PmTransactionTest}.
+ * Test environment projection for {@link io.spine.server.projection.ProjectionTransactionTest}.
  */
-public class TxTestProcessManager
-        extends ProcessManager<ProjectId, Project, Project.Builder> {
+public class TxTestProjection
+        extends Projection<ProjectId, Project, Project.Builder> {
 
     private final List<Message> receivedEvents = newLinkedList();
-    private final @Nullable List<ConstraintViolation> violations;
+    private final List<ConstraintViolation> violations;
 
-    public TxTestProcessManager(ProjectId id) {
+    public TxTestProjection(ProjectId id) {
         this(id, null);
     }
 
-    public TxTestProcessManager(ProjectId id, @Nullable List<ConstraintViolation> violations) {
+    public TxTestProjection(ProjectId id,
+                            @Nullable ImmutableList<ConstraintViolation> violations) {
         super(id);
-        this.violations = violations == null
-                          ? null
-                          : ImmutableList.copyOf(violations);
+        this.violations = violations;
     }
 
     @Override
@@ -64,25 +60,24 @@ public class TxTestProcessManager
         return super.checkEntityState(newState);
     }
 
-    @React
-    Nothing event(PmProjectCreated event) {
+    @Subscribe
+    void event(TxProjectCreated event) {
         receivedEvents.add(event);
         Project newState = Project.newBuilder(state())
                                   .setId(event.getProjectId())
                                   .build();
         builder().mergeFrom(newState);
-        return nothing();
     }
 
     /**
-     * Always throws {@code RuntimeException} to emulate the case of an error in
-     * a reacting method of a Process Manager.
+     * Always throws {@code RuntimeException} to emulate an error in
+     * a subscribing method of a Projection.
      *
-     * @see io.spine.server.procman.PmTransactionTest#createEventThatFailsInHandler()
+     * @see io.spine.server.projection.ProjectionTransactionTest#createEventThatFailsInHandler
      */
-    @React
-    Nothing event(PmTaskAdded event) {
-        throw new RuntimeException("that tests the tx behaviour for process manager");
+    @Subscribe
+    void event(TxTaskAdded e) {
+        throw new RuntimeException("that tests the projection tx behaviour");
     }
 
     public List<Message> receivedEvents() {
