+++ server/src/test/proto/spine/test/tx/tx_entities.proto
@@ -1,0 +1,69 @@
+/*
+ * Copyright 2019, TeamDev. All rights reserved.
+ *
+ * Redistribution and use in source and/or binary forms, with or without
+ * modification, must retain the above copyright notice and the following
+ * disclaimer.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+ * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+ * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+syntax = "proto3";
+
+package spine.test.tx;
+
+import "spine/options.proto";
+
+option (type_url_prefix) = "type.spine.io";
+option java_package = "io.spine.server.entity.given.tx";
+option java_multiple_files = true;
+
+import "google/protobuf/any.proto";
+
+message ProjectId {
+    string id = 1;
+}
+
+enum Status {
+    PS_UNDEFINED = 0;
+    CREATED = 1;
+    STARTED = 2;
+    DONE = 3;
+    CANCELLED = 4;
+}
+
+message Project {
+    option (entity) = {kind: AGGREGATE visibility: FULL};
+
+    // The aggregate ID.
+    //
+    // In order to skip verbose test set up, this ID may be not set.
+    //
+    ProjectId id = 1 [(required) = false];
+
+    string name = 2;
+    repeated Task task = 3;
+    Status status = 4;
+}
+
+message TaskId {
+    int32 id = 1;
+}
+
+message Task {
+    option (entity) = {kind: AGGREGATE visibility: FULL};
+
+    ProjectId id = 1 [(required) = false, (validate) = false];
+    TaskId task_id = 2;
+    string title = 3;
+    string description = 4;
+}
